Fixed the two main freezing bugs. Hurray.

+++ droplet_code/include/scheduler.h
@@ -20,7 +20,7 @@
 typedef struct task
 {
 	uint32_t scheduled_time;
-	uint16_t period;
+	uint32_t period;
 	void (*task_function)(void*);
 	void* arg;
 	struct task* next;
@@ -54,7 +54,7 @@ void task_list_cleanup();
  */
 Task_t* schedule_task(uint32_t time, void (*function)(void*), void* arg);
 // This function primarily calls the above, but always to run 10ms in the future, and then repeat with a certain period.
-Task_t* schedule_periodic_task(uint16_t period, void (*function)(void*), void* arg);
+Task_t* schedule_periodic_task(uint32_t period, void (*function)(void*), void* arg);
 void add_task_to_lsit(Task_t* task);
 void remove_task(Task_t* task); // Removes a task from the queue
 void print_task_queue();
